+++ src/Logic/ProcedureCommonFeatures.php
@@ -12,7 +12,6 @@ declare(strict_types=1);
 
 namespace DemosEurope\DemosplanAddon\XBeteiligung\Logic;
 
-use DemosEurope\DemosplanAddon\Contracts\CurrentContextProviderInterface;
 use DemosEurope\DemosplanAddon\Contracts\Services\CurrentUserProviderInterface;
 use DemosEurope\DemosplanAddon\Contracts\Services\OrgaServiceInterface;
 use DemosEurope\DemosplanAddon\Contracts\Services\ProcedureServiceInterface;
@@ -20,6 +19,7 @@ use DemosEurope\DemosplanAddon\Contracts\Services\ProcedureServiceStorageInterfa
 use DemosEurope\DemosplanAddon\Contracts\Services\ProcedureTypeServiceInterface;
 use DemosEurope\DemosplanAddon\Contracts\Services\TransactionServiceInterface;
 use DemosEurope\DemosplanAddon\Contracts\UserHandlerInterface;
+use DemosEurope\DemosplanAddon\XBeteiligung\Logic\Kommunale\ProcedurePhaseExtractor;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\MessageFactory\KommunaleMessageFactory;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\MessageFactory\PlanfeststellungMessageFactory;
 use DemosEurope\DemosplanAddon\XBeteiligung\Logic\MessageFactory\RaumordnungMessageFactory;
@@ -31,17 +31,18 @@ abstract class ProcedureCommonFeatures
 {
     public function __construct(
         protected readonly CurrentUserProviderInterface       $currentUserProvider,
+        protected readonly EntityManagerInterface             $entityManager,
+        protected readonly KommunaleMessageFactory            $kommunaleMessageFactory,
         protected readonly LoggerInterface                    $logger,
+        protected readonly PlanfeststellungMessageFactory     $planfeststellungMessageFactory,
+        protected readonly ProcedurePhaseExtractor            $procedurePhaseExtractor,
         protected readonly ProcedureServiceInterface          $procedureService,
         protected readonly ProcedureServiceStorageInterface   $procedureServiceStorage,
         protected readonly ProcedureTypeServiceInterface      $procedureTypeService,
-        protected readonly UserHandlerInterface               $userHandler,
-        protected readonly EntityManagerInterface             $entityManager,
-        protected readonly KommunaleMessageFactory            $kommunaleMessageFactory,
         protected readonly RaumordnungMessageFactory          $raumordnungMessageFactory,
-        protected readonly PlanfeststellungMessageFactory     $planfeststellungMessageFactory,
-        protected readonly TranslatorInterface                $translator,
         protected readonly TransactionServiceInterface        $transactionService,
+        protected readonly TranslatorInterface                $translator,
+        protected readonly UserHandlerInterface               $userHandler,
         protected readonly OrgaServiceInterface               $orgaService,
     )
     {